+++ src/modules/fw_pos_control_l1/Landingslope.cpp
@@ -37,7 +37,7 @@
  * @author Thomas Gubler <thomasgubler@gmail.com>
  */
 
-#include "landingslope.h"
+#include "Landingslope.hpp"
 
 #include <mathlib/mathlib.h>
